// src/Models/City/city.ts
export class City{
    SrNo: number;
    Result: number;
    CityCode: string;
    CityName: string;
    CityId: any;

    StateCode: string;
    StateName: string;
    CountryCode: string;
    CountryName: string;

    CreatedBy: string;
    CreatedOn: string;
    UpdatedBy: string;
    UpdatedOn: string;
}
export class Country{
    CountryCode: string;
    CountryName: string;
}

export class Stete{
    StateCode: string;
    StateName: string;
}